LayoutTests/imported/w3c/web-platform-tests/html/semantics/forms/the-selectmenu-element/selectmenu-popup-position-with-zoom.tentative-expected.txt

Custom bottom left

Custom top left
Custom bottom right
Custom top right

PASS The popup should be bottom left positioned
FAIL The popup should be top left positioned assert_equals: expected 0 but got 18
PASS The popup should be bottom right positioned
FAIL The popup should be top right positioned assert_equals: expected 0 but got 18
